清理结构
