+++ src/app/layoutProvider.tsx
@@ -5,17 +5,13 @@ import { NotificationProvider } from "@/context/NotificationProvider";
 import Notification from "@/app/components/Notification";
 import { CSSProperties } from "react";
 import React from "react";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
-
- 
 export  function  LayoutProvider({ children }: { children: React.ReactNode }) {
   const defaultClient = React.useMemo(() => {
     return process.env.NEXT_PUBLIC_IS_MAINNET === "true"
       ? new ccc.ClientPublicMainnet()
       : new ccc.ClientPublicTestnet();
   }, []);
-  
   
   return (
     <ccc.Provider
@@ -46,14 +42,12 @@ export  function  LayoutProvider({ children }: { children: React.ReactNode }) {
         },
       ]}
     >
-      <QueryClientProvider client={new QueryClient()} >
        
         <NotificationProvider>
           {children}
           <Notification />
         </NotificationProvider>
      
-      </QueryClientProvider>
     </ccc.Provider>
   );
 }
